Battle for Kashmir Mechanic
Creates a decision category & set of decisions for India & Pakistan to compete for influence within the provinces of Kashmir as a prelude to the war itself. Having more influence than the other country in a province when the war begins spawns two paramilitary units and grants 10% war support. Both countries can spend 100 political power to gain 100 influence in a province. Pakistan can spend stability to hire private military contractors to gain 50 influence in a province while India can spend 1% party popularity and 500 infantry equipmnt to send in the RSS to gain 50 influence in a province.

todo:
- balancing
- AI decision selection
- other dynamic events to gain/lose influence
- better GFX